fix(ci): avoid testinfra failure due to loss of ssh connection (#1764)
* fix(ci): replace EC2 Instance Connect with cloud-init SSH key injection

Tests are often failing due to the SSH access to the instance.

EC2 Instance Connect push the temporary SSH key which is then available
only for 60 seconds. Recently, errors often occur when the SSH key is
sent to the instance, resulting in a timeout.

We replace runtime SSH key injection via EC2 Instance Connect API with
cloud-init configuration to add the SSH public key during instance initialization.

Note that we are still using EC2 Instance Connect to create the SSH key
pair, but we are not using it to push the key to the instance.

* fix(ci): terminate only the ec2 instance of the matrix job

For the moment, the first matrix job that finishes will terminate all
the ec2 instances running in the current workflow run. This is not what we
want.

This change only terminates the instance that is running the matrix job.

* fix(ci): init.sh completion check should not be blocking

Add optional timeout parameter to run_ssh_command() to check init completion status with a 5-second timeout.