feat: multi version pg_net including 0.19.5 (#1744)
* feat: multi version pg_net including 0.19.5

* feat: bump ver and add tests

* feat: if file exists, run prestart version switch scripts

* tests: bump version to test

* fix: rm unneeded check and fix err codes for file handling

* chore: bump version for testing

* chore: update versions for release

* Add pg-net nixos test in flake checks output

* Move switch_pg_net_version script into separate package

* Fix shellcheck errors

* feat:
limit using verison in extension creation to admin user for all extensions
by using supautils and before-create hooks
add test for this in pg_net

* feat(pg_net): switch version on file system using overlayfs

For pg_net background worker, we need to switch the version of the
extension on the file system (both pg_net.so and pg_net.control need to point
to the desired version).

As the extension is in the Nix store, we cannot simply symlink to the
desired version, as the Nix store is read-only.
To work around this, we use overlayfs to create a writable layer on top
of the pg_net store path.

* Revert "feat:"

This reverts commit 268cadb9eab2b858a376775a9ae086a7be99fe15.

* fix: fmt

* chore: bump version for testin

* fix: control files need module_pathname

* Revert "fix: control files need module_pathname"

This reverts commit 4996d81a8596ffbb903a8eceb2b1c8d0e153bc7c.

This change would not resolve the issue it was targeting.

* chore: version bump for testing

* fix: limit flag to versions less than 0.19.1

* chore: bump version for release

---------

Co-authored-by: Jean-François Roche <jfroche@pyxel.be>