fix(pgmq): replace drop_queue function if exists (#1828)
* fix(pgmq): replace drop_queue function if exists

We introduced a patch to override pgmq.drop_queue(TEXT) to conditionally
drop objects only if they are part of the extension in `ansible/ansible/files/postgresql_extension_custom_scripts/pgmq/after-create.sql`
. This script might have been installed on installations with existing 1.4.4 extensions.

When the user tries to upgrade pgmq from 1.4.4 to 1.5.1, the upgrade
process will fail because the upgrade script doesn't expect
`pgmq.drop_queue(TEXT)` to exist.

This change introduce the a patch to the pgmq extension to use
`CREATE OR REPLACE FUNCTION` instead of `CREATE FUNCTION` for
`pgmq.drop_queue(TEXT)` in the upgrade script from 1.4.5 to 1.5.0.

* feat: test that the pgmq extension works well with the after create script

We create a new custom nixos test for the pgmq extension to test that the extension works
well with the after-create.sql script present in
`ansible/files/postgresql_extension_custom_scripts/pgmq/after-create.sql`.

* chore: bump versions to release post-rebase

---------

Co-authored-by: Sam Rose <samuel@supabase.io>